+++ src/i18n/en/index.js
@@ -156,6 +156,7 @@ export default {
     title: 'Title',
     title_unknown: 'Unknown Title',
     type: 'Type',
+    video_duration: 'Video duration',
     video_title: 'Video title',
     video_url: 'Video URL',
     video: 'Videos',