+++ src/lib/acl-helpers.js
@@ -8,7 +8,7 @@ class ACLHelpers {
       for (let item of results.items) {
         if (item.author.id === context.$store.state.auth.user.uuid) {
           const itemPayload = ObjectUtil.merge({}, payload)
-          itemPayload.uuid = item.uuid
+          itemPayload._uuid = item._uuid
           await context.$store.dispatch(action, itemPayload)
         }
       }
@@ -18,16 +18,16 @@ class ACLHelpers {
   static async updateACL (context) {
     console.debug('setting acl...', context.acl)
     if (this.acl.public) {
-      await this.setACL('acl/set', { role: 'public', uuid: context.timeline.uuid, permissions: ['get'] }, context.acl.recursive)
+      await this.setACL('acl/set', { role: 'public', uuid: context.timeline._uuid, permissions: ['get'] }, context.acl.recursive)
     }
     else {
-      await this.setACL('acl/remove', { role: 'public', uuid: context.timeline.uuid, permission: 'get' }, context.acl.recursive)
+      await this.setACL('acl/remove', { role: 'public', uuid: context.timeline._uuid, permission: 'get' }, context.acl.recursive)
     }
     if (this.acl.group) {
-      await this.setACL('acl/set', { role: context.acl.group, uuid: context.timeline.uuid, permissions: ['get'] }, context.acl.recursive)
+      await this.setACL('acl/set', { role: context.acl.group, uuid: context.timeline._uuid, permissions: ['get'] }, context.acl.recursive)
     }
     if (this.acl.group_remove) {
-      await this.setACL('acl/remove', { role: context.acl.group_remove, uuid: context.timeline.uuid, permission: 'get' }, context.acl.recursive)
+      await this.setACL('acl/remove', { role: context.acl.group_remove, uuid: context.timeline._uuid, permission: 'get' }, context.acl.recursive)
     }
     this.$store.commit('notifications/addMessage', {
       body: 'messages.acl_updated',