+++ src/pages/piecemaker/timelines/sessions.vue
@@ -41,7 +41,6 @@
 
 <script>
   // import axios from 'axios'
-  import { DateTime } from 'luxon'
   import {
     SessionDiagram,
     SessionFilter,
@@ -68,7 +67,7 @@
   //   }
   //   for (let v of videos) {
   //     try {
-  //       const meta = await axios.get(`${process.env.TRANSCODER_HOST}/metadata/${v.annotation.uuid}`, {headers})
+  //       const meta = await axios.get(`${process.env.TRANSCODER_HOST}/metadata/${v.annotation._uuid}`, {headers})
   //       Object.assign(v.meta, meta.data)
   //       videosMeta.push(v)
   //     }
@@ -84,15 +83,15 @@
       SessionStream
     },
     async mounted () {
-      this.map = await this.$store.dispatch('maps/get', this.$route.params.id)
+      this.map = await this.$store.dispatch('maps/get', this.$route.params.uuid)
       this.sessions = await this.generateSessions(this.map.id)
 
       /*
-      this.map = await this.$store.dispatch('maps/get', this.$route.params.id)
+      this.map = await this.$store.dispatch('maps/get', this.$route.params.uuid)
       // const grouped = await this.$store.dispatch('sessions/get', uuid)
       // grouped.sessions = grouped.sessions.map(session => {
-      //   session.start = DateTime.fromISO(session.start)
-      //   session.end = DateTime.fromISO(session.end)
+      //   session.start = session.start
+      //   session.end = session.end
       //   return session
       // })
       const result = await this.$store.dispatch('annotations/find', {
@@ -115,7 +114,7 @@
         })
       const videos = await fetchMetaData(videosBase)
 
-      const millisDist = (3600 || constants.SESSION_DISTANCE_SECONDS) * 1000
+      const millisDist = (3600 || constants.config.SESSION_DISTANCE_SECONDS) * 1000
       const sessions = []
       const defaultSession = { start: undefined, end: undefined, duration: undefined, annotations: [], videos: [] }
       let session = ObjectUtil.merge({}, defaultSession)
@@ -129,7 +128,7 @@
         // TODO: handle other annotations with a duration here
         if (a.body.type === 'Video') {
           video = videos.find(v => {
-            return a.uuid === v.annotation.uuid
+            return a._uuid === v.annotation._uuid
           })
           if (video) {
             if (video.meta && video.meta.duration) {
@@ -155,20 +154,20 @@
 
         if (annotationInside) {
           // only set end if past current session.end
-          if (annotationEnd.toMillis() > session.end.toMillis()) {
+          if (annotationEnd > session.end) {
             session.end = annotationEnd
           }
           session.annotations.push({
             annotation: a,
             duration: annotationDuration,
-            relativeTime: annotationStart.toMillis() - session.start.toMillis(),
+            relativeTime: annotationStart - session.start,
             active: false
           })
           if (video) session.videos.push(video)
         }
         if (!annotationInside || isLastAnnotation) {
           // store current annotation
-          session.duration = session.end.toMillis() - session.start.toMillis()
+          session.duration = session.end - session.start
           if (isNaN(session.duration)) {
             console.error('duration NaN', session)
             session.duration = 0
@@ -215,17 +214,17 @@
         const annotations = result.items.sort(this.$sort.onRef)
         let lastRef, session
         for (let annotation of annotations) {
-          const ref = DateTime.fromISO(annotation.target.selector.value)
-          if (lastRef) console.debug('dist', ref.toMillis() - lastRef.toMillis(), ref.minus(lastRef).toMillis(), maxDist)
-          if (!lastRef || ref.toMillis() - lastRef.toMillis() >= maxDist) {
+          const ref = annotation.target.selector._valueMillis
+          if (lastRef) console.debug('dist', ref - lastRef, maxDist)
+          if (!lastRef || ref - lastRef >= maxDist) {
             session = storeSession(sessions, session)
-            session.start = annotation.target.selector.value
-            session.end = annotation.target.selector.value
+            session.start = annotation.target.selector._valueMillis
+            session.end = annotation.target.selector._valueMillis
           }
-          let sessionEnd = DateTime.fromISO(session.end)
+          let sessionEnd = session.end
           if (lastRef && lastRef > sessionEnd) {
             sessionEnd = lastRef
-            session.end = sessionEnd.toISO()
+            session.end = sessionEnd
           }
           if (annotation.body.type === 'Video') {
             const metadata = await this.$store.dispatch('metadata/get', annotation)
@@ -234,8 +233,8 @@
               annotation
             }
             if (metadata && metadata.duration) {
-              const videoEnd = DateTime.fromISO(annotation.target.selector.value).plus(metadata.duration * 1000)
-              if (videoEnd > ref && videoEnd > sessionEnd) session.end = videoEnd.toISO()
+              const videoEnd = annotation.target.selector._valueMillis + metadata.duration * 1000
+              if (videoEnd > ref && videoEnd > sessionEnd) session.end = videoEnd
             }
             session.videos.push(video)
           }
@@ -281,7 +280,7 @@
         const _this = this
         switch (type) {
         case 'annotate':
-          return _this.$router.push(`/piecemaker/videos/${data.row.uuid}/annotate`)
+          return _this.$router.push(`/piecemaker/videos/${data.row._uuid}/annotate`)
         }
       }
     },